Don't track configured colorscheme
The colorscheme is now set by sourcing an ignored file that sets the
colorscheme as its only action.

What colorschemes are available can be tracked by Vundle as plugins, or
in the plugin/ dir to be tracked by this repository.  But I really don't
want to track what colorscheme has been chosen, especially since it
changes often and makes no real difference.