Add automatic function folding for PHP files
This is intended to automatically create (but not close) folds between
curly brackets.  So folding every function should be a matter just of
zM.  This was taken from something I had for javascript that I'm pretty
sure I got off the internet.  Reading up on how it works is recommended.