Neovim changed the startup file to 'init.vim'
This just renames nvimrc to init.vim, to accomodate the neovim change.

Neovim also changed its startup directory to ~/.config/nvim, so I linked
that to here.

It may still be necessary to deal with the shada file (replacement for
viminfo), and using the XDG variables may be recommended.

From nvim, see :help vim-differences and :help init.vim for more info.